Rename WellKnownTypes to WellKnownTypeSignatures
and add more functional test coverage for string concat of well-known
types, covering all strategies.
