+++ test/langtools/tools/javac/StringConcat/WellKnownTypes.java
@@ -24,96 +24,58 @@
 import com.sun.tools.classfile.*;
 import com.sun.tools.classfile.ConstantPool.*;
 
-import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
-
 /*
  * @test
  * @bug     8273914
  * @summary Indy string concat changes order of operations
  * @modules jdk.jdeps/com.sun.tools.classfile
  *
- * @clean *
  * @compile -XDstringConcat=indy              WellKnownTypes.java
  * @run main WellKnownTypes
  *
- * @clean *
  * @compile -XDstringConcat=indyWithConstants WellKnownTypes.java
+ * @run main WellKnownTypes
+ *
+ * @compile -XDstringConcat=inline            WellKnownTypes.java
  * @run main WellKnownTypes
  */
 
 public class WellKnownTypes {
-    static List<String> actualTypes;
-
     static int idx = 0;
 
     static boolean z = true;
     static char c = (char) 42;
-    static short s = (short) 42;
-    static byte b = (byte) 42;
-    static int i = 42;
-    static long l = 42L;
-    static float f = 42.0f;
-    static double d = 42.0;
+    static byte b = (byte) 43;
+    static short s = (short) 44;
+    static int i = 45;
+    static long l = 46L;
+    static float f = 47.0f;
+    static double d = 48.0;
 
     public static void main(String[] argv) throws Exception {
-        readIndyTypes();
-
-        test("" + WellKnownTypes.class, idx++, "(Ljava/lang/String;)Ljava/lang/String;");
-        test("" + Boolean.valueOf(z), idx++, "(Ljava/lang/Boolean;)Ljava/lang/String;");
-        test("" + Character.valueOf(c), idx++, "(Ljava/lang/Character;)Ljava/lang/String;");
-        test("" + Byte.valueOf(b), idx++, "(Ljava/lang/Byte;)Ljava/lang/String;");
-        test("" + Short.valueOf(s), idx++, "(Ljava/lang/Short;)Ljava/lang/String;");
-        test("" + Integer.valueOf(i), idx++, "(Ljava/lang/Integer;)Ljava/lang/String;");
-        test("" + Long.valueOf(l), idx++, "(Ljava/lang/Long;)Ljava/lang/String;");
-        test("" + Double.valueOf(d), idx++, "(Ljava/lang/Double;)Ljava/lang/String;");
-        test("" + Float.valueOf(f), idx++, "(Ljava/lang/Float;)Ljava/lang/String;");
-        test("" + z, idx++, "(Z)Ljava/lang/String;");
-        test("" + c, idx++, "(C)Ljava/lang/String;");
-        test("" + b, idx++, "(B)Ljava/lang/String;");
-        test("" + s, idx++, "(S)Ljava/lang/String;");
-        test("" + i, idx++, "(I)Ljava/lang/String;");
-        test("" + l, idx++, "(J)Ljava/lang/String;");
-        test("" + d, idx++, "(D)Ljava/lang/String;");
-        test("" + f, idx++, "(F)Ljava/lang/String;");
+        test("" + WellKnownTypes.class, idx++, "class WellKnownTypes");
+        test("" + Boolean.valueOf(z), idx++, "true");
+        test("" + Character.valueOf(c), idx++, "*");
+        test("" + Byte.valueOf(b), idx++, "43");
+        test("" + Short.valueOf(s), idx++, "44");
+        test("" + Integer.valueOf(i), idx++, "45");
+        test("" + Long.valueOf(l), idx++, "46");
+        test("" + Float.valueOf(f), idx++, "47.0");
+        test("" + Double.valueOf(d), idx++, "48.0");
+        test("" + z, idx++, "true");
+        test("" + c, idx++, "*");
+        test("" + b, idx++, "43");
+        test("" + s, idx++, "44");
+        test("" + i, idx++, "45");
+        test("" + l, idx++, "46");
+        test("" + f, idx++, "47.0");
+        test("" + d, idx++, "48.0");
     }
 
-    public static void test(String actual, int index, String expectedType) {
-        String actualType = actualTypes.get(index);
-        if (!actualType.equals(expectedType)) {
-            throw new IllegalStateException(
-                    index
-                            + " Unexpected type: expected = "
-                            + expectedType
-                            + ", actual = "
-                            + actualType);
-        }
-    }
-
-    public static void readIndyTypes() throws Exception {
-        actualTypes = new ArrayList<String>();
-
-        ClassFile classFile =
-                ClassFile.read(
-                        new File(
-                                System.getProperty("test.classes", "."),
-                                WellKnownTypes.class.getName() + ".class"));
-        ConstantPool constantPool = classFile.constant_pool;
-
-        for (Method method : classFile.methods) {
-            if (method.getName(constantPool).equals("main")) {
-                Code_attribute code = (Code_attribute) method.attributes.get(Attribute.Code);
-                for (Instruction i : code.getInstructions()) {
-                    if (i.getOpcode() == Opcode.INVOKEDYNAMIC) {
-                        CONSTANT_InvokeDynamic_info indyInfo =
-                                (CONSTANT_InvokeDynamic_info)
-                                        constantPool.get(i.getUnsignedShort(1));
-                        CONSTANT_NameAndType_info natInfo = indyInfo.getNameAndTypeInfo();
-                        actualTypes.add(natInfo.getType());
-                    }
-                }
-            }
+    public static void test(String actual, int index, String expected) {
+        if (!actual.equals(expected)) {
+      throw new IllegalStateException(
+          index + " Unexpected: expected = " + expected + ", actual = " + actual);
         }
     }
 }
